Added river turning into drain in IPhreeqc version of PHAST
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@8034 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/basicsubs.cpp
@@ -3611,7 +3611,7 @@ basic_callback(double x1, double x2, char * str)
 	}
 	if (basic_fortran_callback_ptr != NULL)
 	{
-		return (*basic_fortran_callback_ptr) (&local_x1, &local_x2, str, strlen(str));
+		return (*basic_fortran_callback_ptr) (&local_x1, &local_x2, str, (int) strlen(str));
 	}
 	return 0;
 }
